refactor: relocate use-layout-breakpoints to canonical src/hooks/
Track T07. src/lib/hooks/ contained a single file (use-layout-breakpoints)
while the other 46 hooks live in src/hooks/. Move the file and update its
four importers (feedback-modal, bottom-nav, navbar, use-bottom-sheet-drag).
The now-empty src/lib/hooks/ directory disappears with the git rename.

No functional change.

Baseline SHA: ff5f24ad29a381b7f0fde8d4bae6028408ef9fd4
